update: OrderList qrCode
